+++ backend/alembic/versions/7d84df7ae218_init_schema_stamp_existing.py
@@ -1,0 +1,26 @@
+"""init schema (stamp existing)
+
+Revision ID: 7d84df7ae218
+Revises: 
+Create Date: 2026-03-22 21:02:03.992729
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '7d84df7ae218'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    pass
+
+
+def downgrade() -> None:
+    pass